tokenomics/pyproject.toml
[tool.poetry]
name = "tokenomics"
version = "0.1.0"
description = ""
authors = ["Pai-Sho <paisho@keeperdao.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
web3 = "^5.30.0"
jsonpickle = "^2.2.0"
pandas = "^1.4.3"
numpy = "^1.23.1"
matplotlib = "^3.5.2"
streamlit = "^1.11.1"
Pillow = "^9.2.0"
python-dotenv = "^0.20.0"
plotly = "^5.9.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
